+++ wemake_python_styleguide/options/defaults.py
@@ -62,3 +62,6 @@ MAX_BASE_CLASSES: Final = 3
 
 #: Maximum number of decorators.
 MAX_DECORATORS: Final = 5
+
+#: Maximum number of same string usage in code.
+MAX_STRING_USAGES: Final = 3