+++ wemake_python_styleguide/types.py
@@ -99,3 +99,4 @@ class ConfigurationOptions(Protocol):
     max_imports: int
     max_base_classes: int
     max_decorators: int
+    max_string_usages: int