+++ wemake_python_styleguide/violations/complexity.py
@@ -41,6 +41,7 @@ Summary
    TooManyElifsViolation
    TooManyForsInComprehensionViolation
    TooManyExceptCasesViolation
+   OverusedStringViolation
 
 Module complexity
 -----------------
@@ -69,6 +70,7 @@ Structures complexity
 .. autoclass:: TooManyElifsViolation
 .. autoclass:: TooManyForsInComprehensionViolation
 .. autoclass:: TooManyExceptCasesViolation
+.. autoclass:: OverusedStringViolation
 
 """
 
@@ -76,6 +78,7 @@ from typing_extensions import final
 
 from wemake_python_styleguide.violations.base import (
     ASTViolation,
+    MaybeASTViolation,
     SimpleViolation,
 )
 
@@ -602,3 +605,30 @@ class TooManyExceptCasesViolation(ASTViolation):
 
     error_template = 'Found too many `except` cases'
     code = 225
+
+
+@final
+class OverusedStringViolation(MaybeASTViolation):
+    """
+    Forbids to over-use string constants.
+
+    Reasoning:
+        When some string is used more than several time in your code,
+        it probably means that this string is a meaningful constant.
+        And should be treated like one.
+
+    Solution:
+        Deduplicate you string usages
+        by defining new functions or constants.
+
+    Configuration:
+        This rule is configurable with ``--max-string-usages``.
+        Default:
+        :str:`wemake_python_styleguide.options.defaults.MAX_STRING_USAGES`
+
+    .. versionadded:: 0.10.0
+
+    """
+
+    error_template = 'Found string constant over-use: {0}'
+    code = 226