+++ wemake_python_styleguide/visitors/ast/modules.py
@@ -7,8 +7,8 @@ from typing_extensions import final
 
 from wemake_python_styleguide import constants
 from wemake_python_styleguide.logics.filenames import get_stem
-from wemake_python_styleguide.logics.nodes import get_context, is_doc_string
 from wemake_python_styleguide.logics.naming.constants import is_constant
+from wemake_python_styleguide.logics.nodes import get_context, is_doc_string
 from wemake_python_styleguide.types import AnyNodes
 from wemake_python_styleguide.violations.best_practices import (
     BadMagicModuleFunctionViolation,