+++ CHANGELOG.md
@@ -8,7 +8,9 @@ We used to have incremental versioning before `0.1.0`.
 
 ### Features
 
+- Validates options that are passed with `flake8`
 - Forbids to use module level mutable constants
+- Forbids to over-use strings
 
 
 ## 0.9.0